Correctly find the difference between the peaks.
Did it in the wrong place, I thought a method looped through the peaks, but it looped through each sample of a path. This was because I'd originally named it plural. Now I do the functionality after that method has ran for each path, and all peaks are found. Seems to be coping with appendData of my original big double values so I've removed the divide by 100, dunno why it's working now.