Mark maxima and minima on the photo, attempt to match up horizons
Major fix to take into account that the coordinates saved were
compressed to be only the width of the number of points, not the width
of the photo. This was to save space, allowing for a much smaller list.

Next: test the accuracy of the value got of the accuracy of the matching