Now works alright with all images that have peaks
Adjust weights to work with the majority of photos
Check gradient from EVERY pixel, not every *search width* pixel to avoid
missing any peaks/troughs.
When adding pointy peaks, it checked if the gradient before was significant
but not the gradient after (just checked if +ve or -ve), now it does.